✨ Adding a way to get the info menu
It is possible to get the info menu by the getInfoMenu method. It returns a menu with each info associated with an item with a name. These items are configurable in config by specifying their Material. The money item and the xP item have their data directly inside their name.