✨ Make the items of the challenges separate
When the items of the challenges are merged, PaperMc retrieve the challenge in config and set the amount of an item to 0, set his type to air and his damage to -1 if the amount of the item is greater than 64.
I think that PaperMc do this because of the optimization of this minecraft server, so it does not like when there is more than 64 items in a stack.